requirements.txt
python-dotenv
streamlit
pandas
mysql-connector-python